Acoustic

Acoustic

Acoustic type and functions.

MultipleScattering.Acoustic — Type.

Acoustic{T<:AbstractFloat,Dim}(ρ::T, c::Complex{T})
Acoustic(ρ::T, c::Union{T,Complex{AbstractFloat}}, Dim::Integer)

Physical properties for a homogenous isotropic acoustic medium with wavespeed (c) and density (ρ)

Simulations in this medium produce scalar (1D) fields in Dim dimensions.

source

MultipleScattering.hard — Method.

hard(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.impedance — Method.

impedance(medium::Acoustic)

Characteristic specific acoustic impedance (z₀) of medium

source

MultipleScattering.pressure_release — Method.

pressure_release(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.regular_basis_function — Method.

Basis function when inside a particle. Assumes particle is a circle, which approximately works for all shapes.

source

MultipleScattering.rigid — Method.

rigid(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.soft — Method.

soft(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.sound_hard — Method.

sound_hard([T::Type = Float64,] Dim::Integer)

Construct physical properties of a sound hard acoustic object with type T and dimension Dim. Also known as rigid and equivalent to a zero_neumann pressure boundary condition.

source

MultipleScattering.sound_soft — Method.

sound_soft([T::Type = Float64,] Dim::Integer)

Construct physical properties of a sound hard acoustic object with type T and dimension Dim. Equivalent to a zero_dirichlet pressure boundary condition.

source

MultipleScattering.zero_dirichlet — Method.

zero_dirichlet(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.zero_neumann — Method.

zero_neumann(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.internal_field — Method.

internal_field(x::SVector{2,T}, p::Particle{T,2,Acoustic{T,2},Circle{T}}, sim::FrequencySimulation{T,2,Acoustic{T,2}}, ω::T, scattering_coefficients::AbstractVector{Complex{T}})

The internal field for a 2D circlular acoustic particle in a 2D acoustic medium.

source

MultipleScattering.t_matrix — Method.

t_matrix(Particle{T,2,Acoustic{T,2},Circle{T}}, Acoustic{T,2}, ω, order)

The T-matrix for a 2D circlular acoustic particle in a 2D acoustic medium.

source

MultipleScattering.t_matrix — Method.

t_matrix(CapsuleParticle{T,2,Acoustic{T,2},Circle{T}}, Acoustic{T,2}, ω, order)

The T-matrix for a 2D circlular capsule particle in an acoustic medium.

source

MultipleScattering.plane_source — Method.

plane_source(medium::Acoustic, source_position, source_direction=[1,0], amplitude=1)::Source{Acoustic}

Create 2D Acoustic planar wave Source

source

MultipleScattering.point_source — Method.

point_source(medium::Acoustic, source_position, amplitude=1)::Source{Acoustic}

Create 2D Acoustic point Source (zeroth Hankel function of first type)

source