Random

Random

Random particles

MultipleScattering.random_particles — Function.

random_particles(particle_medium, particle_shape, box_shape, volume_fraction::Number;
    seed=Base.Random.make_seed())
random_particles(particle_medium, particle_shape, box_shape, N::Integer;
    seed=Base.Random.make_seed())

Generate N random particles that fit inside box_shape (or fill with volume_fraction)

Specify seed to make output deterministic. Algorithm places particles unifomly randomly inside the bounding rectangle of box_shape and discards particle if it overlaps (based on outer radius) or does not lies completely in box box.

source

Moments

MultipleScattering.calculate_moments — Function.

calculate_moments(results, n; applytofield=real)::Vector{Matrix}

Calculate moments up to n of results at each position and wavenumber/time, after applying applytofield.

source