Acoustic

Acoustic

Acoustic type and functions.

MultipleScattering.Acoustic — Type.

Acoustic{T<:AbstractFloat,Dim}(ρ::T, c::Complex{T})
Acoustic(ρ::T, c::Union{T,Complex{AbstractFloat}}, Dim::Integer)

Physical properties for a homogenous isotropic acoustic medium with wavespeed (c) and density (ρ)

Simulations in this medium produce scalar (1D) fields in Dim dimensions.

source

MultipleScattering.impedance — Function.

impedance(medium::Acoustic)

Characteristic specific acoustic impedance (z₀) of medium

source

MultipleScattering.sound_hard — Function.

sound_hard([T::Type = Float64,] Dim::Integer)

Construct physical properties of a sound hard acoustic object with type T and dimension Dim. Also known as rigid and equivalent to a zero_neumann pressure boundary condition.

source

MultipleScattering.hard — Function.

hard(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.rigid — Function.

rigid(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.zero_neumann — Function.

zero_neumann(host_medium::Acoustic)

See sound_hard.

source

MultipleScattering.sound_soft — Function.

sound_soft([T::Type = Float64,] Dim::Integer)

Construct physical properties of a sound hard acoustic object with type T and dimension Dim. Equivalent to a zero_dirichlet pressure boundary condition.

source

MultipleScattering.soft — Function.

soft(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.pressure_release — Function.

pressure_release(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.zero_dirichlet — Function.

zero_dirichlet(host_medium::Acoustic)

See sound_soft.

source

MultipleScattering.plane_source — Function.

plane_source(medium::Acoustic, source_position, source_direction=[1,0], amplitude=1)::Source{Acoustic}

Create 2D Acoustic planar wave Source

source

MultipleScattering.point_source — Function.

point_source(medium::Acoustic, source_position, amplitude=1)::Source{Acoustic}

Create 2D Acoustic point Source (zeroth Hankel function of first type)

source