Base

Base

Base

Shapes

Physical properties

Particles

Source

Simulation

Shapes

Shape types and functions.

MultipleScattering.Shape — Type.

Abstract idea which defines the external boundary of object.

source

MultipleScattering.origin — Function.

origin(shape::Shape)::SVector

Origin of shape, typically the center

source

MultipleScattering.iscongruent — Method.

iscongruent(p1::Shape, p2::Shape)::Bool
≅(p1::Shape, p2::Shape)::Bool

True if shapes are the same but in different positions (origins), standard mathematical definition.

source

MultipleScattering.congruent — Function.

congruent(s::Shape, x)::Shape

Create shape congruent to s but with origin at x

source

MultipleScattering.bounding_rectangle — Function.

Returns rectangle which completely encloses the shapes

source

MultipleScattering.boundary_functions — Function.

volume(shape::Shape)::NTuple{Function,Dim)

Returns Tuple of Dim Functions which define outer boundary of shape when given boundary coordinate t∈[0,1]

source

MultipleScattering.name — Function.

name(shape::Shape)::String

Name of a shape

source

MultipleScattering.outer_radius — Function.

outer_radius(shape::Shape{T})::T

The radius of a circle which completely contains the shape

source

MultipleScattering.volume — Function.

volume(shape::Shape{T})::T

Volume of a shape

source

MultipleScattering.Circle — Type.

Circle([origin=zeros(),] radius)

2D Shape where boundary is a fixed distance from the origin.

source

MultipleScattering.Rectangle — Type.

Rectangle([origin::AbstractVector{T}=zeros(),] width::T, Height::T)
Rectangle(bottomleft::AbstractVector{T}, topright::AbstractVector{T})

2D Shape with axis aligned sides, defined by width, height and origin (at the center).

source

MultipleScattering.TimeOfFlightFromPoint — Type.

A shape where anything inside could cause a disturbance at the listener position from a point source wavefront starting at the listener. Also everything inside has a positive x coordinate. It is equivalent to a segment of a circle.

More precisely, if the listener is at (l_x,l_y) then the interior of the shape is defined as sqrt((x-l_x)^2+(y-l_y)^2)<time and x>0

source

MultipleScattering.TimeOfFlight — Type.

A shape where anything inside could cause a disturbance at the listener position from a planar wavefront parallel to the y axis starting at the listener. Also everything inside has a positive x coordinate.

More precisely, if the listener is at (l_x,l_y) then the interior of the shape is defined as x-l_x+sqrt((x-l_x)^2+(y-l_y)^2)<time and x>0

source

MultipleScattering.Sphere — Type.

Sphere([origin=zeros(),] radius)

3D Shape where boundary is a fixed distance from the origin.

source

Physical properties

Physical properties types and functions.

MultipleScattering.PhysicalProperties — Type.

Holds information about the physical properties of the medium, the dimension of the field and the number of dimensions it is a function of.

source

MultipleScattering.field_dim — Function.

Extract the dimension of the field of this physical property

source

Extract the dimension of the field of this type of physical property

source

MultipleScattering.dim — Function.

Extract the dimension of the space that this physical property lives in

source

Extract the dimension of the space that this type of physical property lives in

source

MultipleScattering.basis_function — Function.

Basis functions in a specific dimension for a specific physics type.

source

Basis function when inside a particle. Assumes particle is a circle, which approximately works for all shapes.

source

MultipleScattering.internal_field — Function.

the field inside an AbstractParticle a some given point x.

source

MultipleScattering.boundary_data — Function.

A tuples of vectors of the field close to the boundary of the shape. The field is calculated from sim::FrequencySimulation, but the PhysicalProperties inside and outside of the shape are assumed to be given by inside_medium and outside_medium.

source

Particles

Particle types and functions.

MultipleScattering.AbstractParticle — Type.

Object we can scatter waves off

Subtypes will contain information about shape and material properties. Most crucially, they will implement the t_matrix function

source

MultipleScattering.Particle — Type.

Particle(medium::PhysicalProperties, shape::Shape)

Create particle with inner medium and shape (types and dimension must agree).

source

MultipleScattering.CapsuleParticle — Type.

CapsuleParticle(outer::Particle, inner::Particle)

A particle within another particle, both with the same shape type and origin.

source

MultipleScattering.iscongruent — Method.

iscongruent(p1::AbstractParticle, p2::AbstractParticle)::Bool
≅(p1::AbstractParticle, p2::AbstractParticle)::Bool

Returns true if medium and shape of particles are the same, ignoring origin, false otherwise.

source

Source

Source types and functions.

MultipleScattering.Source — Type.

Source{P,T}(field::Function,coef::Function) where P<:PhysicalProperties{T,Dim,FieldDim}

A struct which describes the source that drives/forces the whole system. It is also described as an incident wave.

x = [1.0,0.0]
ω = 1.0
Source.field(x,ω)

should return the wave field at position 'x' and angular frequency 'ω'

source

MultipleScattering.self_test — Function.

Check that the source functions return the correct types

source

Check that the discrete impulse vectors are the right sizes

source

Check that the continuous impulse functions return the correct types

source

Simulation

Simulation types and functions.

MultipleScattering.FrequencySimulation — Type.

FrequencySimulation(medium::PhysicalProperties,
                    [particles::AbstractParticles=[],]
                    source::Source)

Build a FrequencySimulation. If particles are not provided, an empty array is used.

After building, you can run the simulation to get a FrequencySimulationResult.

source

Base.run — Method.

run(sim::FrequencySimulation, x, ω; basis_order=5)

Run the simulation sim for the position x and angular frequency ω.

Position can be an SVector or Vector{SVector} and frequency can be a float or vector of floats.

source

Base.run — Method.

run(sim::FrequencySimulation, rectangle;
    res=20, xres=res, yres=res, basis_order=5)

Run the simulation sim for a grid of positions in rectangle and for angular frequency ω.

Frequency can be a float or vector of floats. The resolution of the grid points is defined by xres and yres.

source

MultipleScattering.FrequencySimulationResult — Type.

Struct to hold results of a FrequencySimulation

source

MultipleScattering.forcing — Function.

forcing(source::Source, particles::AbstractParticles, ω::AbstractFloat, Nh::Integer)::Vector{Complex}

Create forcing vector from source, forms the right hand side of matrix equation to find basis_coefficients.

source

MultipleScattering.basis_coefficients — Function.

basis_coefficients(sim::FrequencySimulation, ω::AbstractFloat; basis_order::Int=5)::Matrix{Complex}

Return coefficients for bases around each particle for a given simulation and angular frequency (ω).

source

MultipleScattering.field — Function.

field(result::SimulationResult, [i::Integer, j::Integer])

Get field from result, optionally specifying indices.

Returns single value of/matrix of complex SVectors() if vector field, and complex float if scalar field.

source

MultipleScattering.scattering_matrix — Function.

Create the matrix S which will be inverted to find the scattering coefficients. Currently assumes 2D.

source

MultipleScattering.t_matrix — Function.

Returns a 2M+1 by 2M+1 T-matrix for particle with specific shape, physical properties in a medium with a specific physical property at a specific angular wavenumber. See doc/T-matrix.pdf for details.

source

MultipleScattering.get_t_matrices — Function.

Returns vector of T-matrices from a vector of particles in a specific domain. Can save computation if multiple of the same kind of particle are present in the vector.

source