Base

Base

Base

Shape

Physical property

Particle

Source

Impulse

Simulation

Shape

Shape types and functions.

MultipleScattering.Shape — Type.

Abstract idea which defines the external boundary of object.

source

MultipleScattering.origin — Function.

origin(shape::Shape)::SVector

Origin of shape, typically the center

source

MultipleScattering.iscongruent — Method.

iscongruent(p1::Shape, p2::Shape)::Bool
≅(p1::Shape, p2::Shape)::Bool

True if shapes are the same but in different positions (origins), standard mathematical definition.

source

MultipleScattering.congruent — Function.

congruent(s::Shape, x)::Shape

Create shape congruent to s but with origin at x

source

MultipleScattering.bounding_rectangle — Function.

Returns rectangle which completely encloses the shapes

source

MultipleScattering.boundary_functions — Function.

volume(shape::Shape)::NTuple{Function,Dim)

Returns Tuple of Dim Functions which define outer boundary of shape when given boundary coordinate t∈[0,1]

source

MultipleScattering.name — Function.

name(shape::Shape)::String

Name of a shape

source

MultipleScattering.outer_radius — Function.

outer_radius(shape::Shape{T})::T

The radius of a circle which completely contains the shape

source

MultipleScattering.volume — Function.

volume(shape::Shape{T})::T

Volume of a shape

source

MultipleScattering.Circle — Type.

Circle([origin=zeros(),] radius)

2D Shape where boundary is a fixed distance from the origin.

source

MultipleScattering.Rectangle — Type.

Rectangle([origin::AbstractVector{T}=zeros(),] width::T, Height::T)
Rectangle(bottomleft::AbstractVector{T}, topright::AbstractVector{T})

2D Shape with axis aligned sides, defined by width, height and origin (at the center).

source

MultipleScattering.TimeOfFlightFromPoint — Type.

A shape where anything inside could cause a disturbance at the listener position from a point source wavefront starting at the listener. Also everything inside has a positive x coordinate. It is equivalent to a segment of a circle.

More precisely, if the listener is at (lx,ly) then the interior of the shape is defined as sqrt((x-lx)^2+(y-ly)^2)<time and x>0

source

MultipleScattering.TimeOfFlight — Type.

A shape where anything inside could cause a disturbance at the listener position from a planar wavefront parallel to the y axis starting at the listener. Also everything inside has a positive x coordinate.

More precisely, if the listener is at (lx,ly) then the interior of the shape is defined as x-lx+sqrt((x-lx)^2+(y-l_y)^2)<time and x>0

source

MultipleScattering.Sphere — Type.

Sphere([origin=zeros(),] radius)

3D Shape where boundary is a fixed distance from the origin.

source

Physical property

Physical properties types and functions.

MultipleScattering.PhysicalMedium — Type.

Holds information about the physical properties of the medium, the dimension of the field and the number of dimensions it is a function of.

source

Missing docstring.

Missing docstring for field_dim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dim. Check Documenter's build log for details.

MultipleScattering.outgoing_basis_function — Function.

Basis of outgoing wave functions in a specific dimension for specific physcial properties..

source

MultipleScattering.regular_basis_function — Function.

Basis functions in a specific dimension for specific physcial properties..

source

Basis function when inside a particle. Assumes particle is a circle, which approximately works for all shapes.

source

MultipleScattering.internal_field — Function.

the field inside an AbstractParticle a some given point x.

source

MultipleScattering.boundary_data — Function.

A tuples of vectors of the field close to the boundary of the shape. The field is calculated from sim::FrequencySimulation, but the PhysicalMedium inside and outside of the shape are assumed to be given by insidemedium and outsidemedium.

source

Particle

Particle types and functions.

MultipleScattering.AbstractParticle — Type.

Object we can scatter waves off

Subtypes will contain information about shape and material properties. Most crucially, they will implement the t_matrix function

source

MultipleScattering.Particle — Type.

Particle(medium::PhysicalMedium, shape::Shape)

Create particle with inner medium and shape (types and dimension must agree).

source

MultipleScattering.CapsuleParticle — Type.

CapsuleParticle(outer::Particle, inner::Particle)

A particle within another particle, both with the same shape type and origin.

source

MultipleScattering.iscongruent — Method.

iscongruent(p1::AbstractParticle, p2::AbstractParticle)::Bool
≅(p1::AbstractParticle, p2::AbstractParticle)::Bool

Returns true if medium and shape of particles are the same, ignoring origin, false otherwise.

source

Source

Source types and functions.

MultipleScattering.Source — Type.

Source(medium::P, field::Function, coef::Function)

Is a struct type which describes the source field that drives/forces the whole system. It is also described as an incident wave. It has three fields Source.medium, Source.field, and Source.coef.

The source field at the position 'x' and angular frequency 'ω' is given by

x = [1.0,0.0]
ω = 1.0
Source.field(x,ω)

The field Source.coef regularbasisfunction(medium::Acoustic{T,2}, ω::T)

source

MultipleScattering.self_test — Method.

Check that the source functions return the correct types

source

MultipleScattering.source_expand — Method.

source_expand(Source, centre; basis_order = 4)

Returns a function of (x,ω) which approximates the value of the source at (x,ω). That is, the source is written in terms of a regular basis expansion centred at centre.

source

MultipleScattering.AbstractSource — Type.

Represent any source (incident) wave

Subtypes may have a symmetry (such as PlaneSource) and will contain information about physical medium.

source

MultipleScattering.PlaneSource — Type.

PlaneSource(medium::P, field::Function, coef::Function)

Is a struct type which describes a plane-wave source that drives/forces the whole system. It has four fields: a polar angle θ, an azimuth angle φ, a physical medium, and an amplitude vector.

source

Impulse

Impulse types and functions.

MultipleScattering.ContinuousImpulse — Type.

See also: DiscreteImpulse, frequency_to_time

ContinuousImpulse{T<:AbstractFloat}

A struct used to represnt analytic impulse functions. Has two fields: in_time a function of time t, and in_freq a function of the angular frequency ω. in_freq should be the Fourier transform of in_time, though this is not enforced.

We use the Fourier transform convention: F(ω) = ∫ f(t)exp(imωt) dt, f(t) = (2π)^(-1) * ∫ F(ω)exp(-imωt) dω.

An impluse f(t) is convoluted in time with the field u(t), however we avoid the convlution by working with the fourier transform F(ω) of the impulse f(t), which results in

frequency to time: (2π)^(-1) * ∫ F(ω)U(ω)exp(-imωt) dω

source

MultipleScattering.DiscreteImpulse — Type.

See also: ContinuousImpulse, frequency_to_time, DiscreteGaussianImpulse

DiscreteImpulse{T<:AbstractFloat}

A struct used to represent a numerical impulse. Only the fields: in_freq which is the frequency response vector, and the frequency vector ω are required to use this struct to use in the function frequency_to_time.

source

MultipleScattering.DiscreteGaussianImpulse — Method.

See also: ContinuousImpulse, TimeDiracImpulse

DiscreteGaussianImpulse(t_vec[, ω_vec])

Returns a discretised gaussian impulse.

source

MultipleScattering.FreqDiracImpulse — Method.

Dirac Delta function of unit area in the frequency domain centred at ω=ω0.

Warning: in frequency space this is a singuarity and so may lead to unexpected behaviour.

source

MultipleScattering.GaussianImpulse — Method.

See also: ContinuousImpulse, TimeDiracImpulse

GaussianImpulse(maxω[, a = 3.0/maxω^2])

Returns a gaussian impulse function, which in the frequency domain is exp(-aω^2)(2sqrt(a*pi)).

source

MultipleScattering.TimeDiracImpulse — Method.

TimeDiracImpulse(t0::T)

Dirac Delta function of unit area in the time domain centred at t=t0.

Warning: in the time domain this is a singuarity and so may lead to unexpected behaviour.

source

MultipleScattering.self_test — Method.

Check that the continuous impulse functions return the correct types

source

MultipleScattering.self_test — Method.

Check that the discrete impulse vectors are the right sizes

source

MultipleScattering.firstnonzero — Method.

Returns the first element of array which isn't zero (assumes elements are increasing and distinct)

source

MultipleScattering.frequency_to_time — Method.

Convert a FrequencySimulationResult into a TimeSimulationResult by using the inverse fourier transform. Assumes only positive frequencies and a real time signal

source

MultipleScattering.frequency_to_time — Method.

See also: DiscreteImpulse, ContinuousImpulse

Calculates the time response from the frequency response by approximating an inverse Fourier transform. The time signal is assumed to be real and the frequenices ω_vec are assumed to be positive (can include zero) and sorted. The result is convoluted in time ωith the user specified impulse.

We use the Fourier transform convention: F(ω) = ∫ f(t)exp(imωt) dt f(t) = (2π)^(-1) * ∫ F(ω)exp(-imωt) dt

To easily sample any time, the default is not FFT, but a discrete version of the transform above.

source

MultipleScattering.t_to_ω — Method.

The inverse of ωtot if ω_vec[1] == 0

source

MultipleScattering.time_to_frequency — Method.

Convert a TimeSimulationResult into a FrequencySimulationResult by using the fourier transform. Assumes only positive frequencies and a real time signal

source

MultipleScattering.time_to_frequency — Method.

The inverse of the function frequencytotime (only an exact inverse when using :dft integration). We use the Fourier transform convention: F(ω) = ∫ f(t)exp(imω*t) dt

source

MultipleScattering.ω_to_t — Method.

returns an array of time from the frequency array ωvec. Uses the same convention for sampling the time as the discrete Fourier transfrom. Assumes ωvec is ordered and non-negative.

source

Simulation

Simulation types and functions.

MultipleScattering.FrequencySimulation — Type.

FrequencySimulation([particles::AbstractParticles=[],]
                    source::Source)

Build a FrequencySimulation. If particles are not provided, an empty array is used.

After building, you can run the simulation to get a FrequencySimulationResult.

source

Base.run — Method.

run(sim::FrequencySimulation, x, ω; basis_order=5)

Run the simulation sim for the position x and angular frequency ω.

Position can be an SVector or Vector{SVector} and frequency can be a float or vector of floats.

source

Base.run — Method.

run(sim::FrequencySimulation, rectangle;
    res=20, xres=res, yres=res, basis_order=5)

Run the simulation sim for a grid of positions in rectangle and for angular frequency ω.

Frequency can be a float or vector of floats. The resolution of the grid points is defined by xres and yres.

source

MultipleScattering.FrequencySimulationResult — Type.

Struct to hold results of a FrequencySimulation

source

MultipleScattering.basis_coefficients — Function.

basis_coefficients(sim::FrequencySimulation, ω::AbstractFloat; basis_order::Int=5)::Matrix{Complex}

Return coefficients for bases around each particle for a given simulation and angular frequency (ω).

source

MultipleScattering.field — Function.

field(result::SimulationResult, [i::Integer, j::Integer])

Get field from result, optionally specifying indices.

Returns single value of/matrix of complex SVectors() if vector field, and complex float if scalar field.

source

MultipleScattering.scattering_matrix — Function.

Create the matrix S which will be inverted to find the scattering coefficients. Currently assumes 2D.

source

MultipleScattering.t_matrix — Function.

t_matrix(particle, medium, ω, order)

Returns a finite T-matrix, with size depending on order, for a specific particle within a medium with specific physical properties.

source

t_matrix(Particle{T,2,Acoustic{T,2},Circle{T}}, Acoustic{T,2}, ω, order)

The T-matrix for a 2D circlular acoustic particle in a 2D acoustic medium.

source

t_matrix(CapsuleParticle{T,2,Acoustic{T,2},Circle{T}}, Acoustic{T,2}, ω, order)

The T-matrix for a 2D circlular capsule particle in an acoustic medium.

source

MultipleScattering.get_t_matrices — Function.

Returns vector of T-matrices from a vector of particles in a specific domain. Can save computation if multiple of the same kind of particle are present in the vector.

source